requirements.lock.txt
aiofiles==23.2.1
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1737352602016/work
attrs==25.4.0
beautifulsoup4==4.14.3
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1749686179973/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1756599213019/work
cellpose==4.0.7
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1759648874697/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1761202952988/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
click==8.3.1
cloudpickle==3.1.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1760108736307/work/conda-src
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1745834476052/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1736345463896/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1751548120229/work
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1734729835256/work
et_xmlfile==2.0.0
fastapi==0.123.0
fastremap==1.17.7
ffmpy==1.0.0
filelock==3.20.0
fill_voids==2.1.1
fonttools==4.61.0
frozendict @ file:///Users/runner/miniforge3/conda-bld/frozendict_1756048002891/work
fsspec==2025.10.0
gradio==3.50.2
gradio_client==0.6.1
graphviz==0.21
greenlet==3.3.0
groovy==0.1.2
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hf-xet==1.2.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
imagecodecs==2025.11.11
ImageIO==2.37.2
importlib_resources==6.5.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1733814567314/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1756754188614/work
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==1.1.3
langchain-core==1.2.0
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
lazy_loader==0.4
libmambapy @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_libmambapy_1760967438/work/libmambapy
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mcp==1.23.1
mdurl==0.1.2
menuinst @ file:///Users/runner/miniforge3/conda-bld/menuinst_1761299914766/work
mpmath==1.3.0
narwhals==2.13.0
natsort==8.4.0
networkx==3.6
numba==0.63.1
numpy==2.2.6
openai==2.13.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
orjson==3.11.4
ormsgpack==1.12.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.3
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1759953252/work
playwright==1.57.0
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1747339660894/work
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1757744682161/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.6
pyparsing==3.2.5
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1755614211359/work
requests-toolbelt==1.0.0
rich==14.2.0
roifile==2025.5.10
rpds-py==0.30.0
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1761160761767/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1760564263846/work
ruff==0.14.7
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
segment-anything==1.0
semantic-version==2.10.0
sentence-transformers==5.1.2
setuptools==80.9.0
shap==0.50.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
soupsieve==2.8.1
sse-starlette==3.0.3
starlette==0.50.0
svgutils==0.3.4
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.10.16
tokenizers==0.22.1
tomlkit==0.12.0
torch==2.9.1
torchvision==0.24.1
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
transformers==4.57.3
truststore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_truststore_1753886790/work
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
uuid_utils==0.12.0
uvicorn==0.38.0
websockets==11.0.3
wheel==0.45.1
xgboost==3.1.2
xlrd==2.0.2
xxhash==3.6.0
zstandard==0.25.0
